Move Cloud Foundry-specific health endpoint tests to separate class
This avoids relying on Cloud Foundry exposure logic in
HealthEndpointAutoConfigurationTests.

See gh-45721
